requirements.txt
openai==1.3.0
pydantic==2.5.0
rich==13.6.0
python-dotenv==1.0.0
jsonpickle==3.0.2
httpx==0.23.0
flask==2.3.3
flask-cors==4.0.0 